feat(firmware): implement Tier 1 critical safety and reliability features
- Add NVS state persistence (Preferences.h) with settled debounce wear leveling for dimmable channels and configurable boot restoration policy
- Implement Central Task Watchdog Timer (WDT) with 10s timeout and FreeRTOS packet queue to eliminate thread safety violations between wifi_task and loopTask
- Implement RFC 6479 64-packet sliding window anti-replay filter with 32-bit monotonic sequence counters in SwitchMessage
- Add configurable auto-off safety timer (10 min) to DigitalChannel for Channel 4 (Water Pump) with boot-off safety constraint
- Add mock Preferences support in firmware/lib/arduino_mock/ and comprehensive native unit tests verifying all features
- Update docs/roadmap.md marking Tier 1 features as completed

+++ firmware/lib/protocol/protocol.h
@@ -18,7 +18,7 @@ struct __attribute__((packed)) SwitchMessage {
     uint8_t action;          // ActionType cast to uint8_t
     int8_t rotation_steps;   // Encoder relative rotation step increment/decrement
     float battery_voltage;   // Voltage of remote battery for power monitoring
-    uint16_t seq;            // Sequence number for replay protection
+    uint32_t seq;            // 32-bit monotonic sequence number for anti-replay
 };
 
 #endif // PROTOCOL_H
